Move era agnostic key based commands out of the EraBased folder
