+++ cardano-cli/cardano-cli.cabal
@@ -62,6 +62,7 @@ library
     Cardano.CLI.Commands.Debug.LogEpochState
     Cardano.CLI.Commands.Debug.TransactionView
     Cardano.CLI.Commands.Hash
+    Cardano.CLI.Commands.Key
     Cardano.CLI.Commands.Node
     Cardano.CLI.Commands.Ping
     Cardano.CLI.Environment
@@ -74,7 +75,6 @@ library
     Cardano.CLI.EraBased.Commands.Governance.DRep
     Cardano.CLI.EraBased.Commands.Governance.Poll
     Cardano.CLI.EraBased.Commands.Governance.Vote
-    Cardano.CLI.EraBased.Commands.Key
     Cardano.CLI.EraBased.Commands.Query
     Cardano.CLI.EraBased.Commands.StakeAddress
     Cardano.CLI.EraBased.Commands.StakePool
@@ -89,7 +89,6 @@ library
     Cardano.CLI.EraBased.Options.Governance.DRep
     Cardano.CLI.EraBased.Options.Governance.Poll
     Cardano.CLI.EraBased.Options.Governance.Vote
-    Cardano.CLI.EraBased.Options.Key
     Cardano.CLI.EraBased.Options.Query
     Cardano.CLI.EraBased.Options.StakeAddress
     Cardano.CLI.EraBased.Options.StakePool
@@ -108,7 +107,6 @@ library
     Cardano.CLI.EraBased.Run.Governance.GenesisKeyDelegationCertificate
     Cardano.CLI.EraBased.Run.Governance.Poll
     Cardano.CLI.EraBased.Run.Governance.Vote
-    Cardano.CLI.EraBased.Run.Key
     Cardano.CLI.EraBased.Run.Query
     Cardano.CLI.EraBased.Run.StakeAddress
     Cardano.CLI.EraBased.Run.StakePool
@@ -129,6 +127,7 @@ library
     Cardano.CLI.Options
     Cardano.CLI.Options.Debug
     Cardano.CLI.Options.Hash
+    Cardano.CLI.Options.Key
     Cardano.CLI.Options.Node
     Cardano.CLI.Options.Ping
     Cardano.CLI.Orphans
@@ -141,6 +140,7 @@ library
     Cardano.CLI.Run.Debug.LogEpochState
     Cardano.CLI.Run.Debug.TransactionView
     Cardano.CLI.Run.Hash
+    Cardano.CLI.Run.Key
     Cardano.CLI.Run.Node
     Cardano.CLI.Run.Ping
     Cardano.CLI.TopHandler
