+++ cardano-cli/src/Cardano/CLI/Commands.hs
@@ -8,6 +8,7 @@ where
 import           Cardano.CLI.Byron.Commands (ByronCommand)
 import           Cardano.CLI.Commands.Debug
 import           Cardano.CLI.Commands.Hash (HashCmds)
+import           Cardano.CLI.Commands.Key
 import           Cardano.CLI.Commands.Node
 import           Cardano.CLI.Commands.Ping (PingCmd (..))
 import           Cardano.CLI.EraBased.Commands
@@ -20,8 +21,10 @@ data ClientCommand
   = AnyEraCommand AnyEraCommand
   | -- | Byron Related Commands
     ByronCommand ByronCommand
-  | -- | Era-agnostic hashing commands
+  | -- | Era agnostic hashing commands
     HashCmds HashCmds
+  | -- | Era agnostic key commands
+    KeyCommands KeyCmds
   | -- | Era agnostic node commands
     NodeCommands NodeCmds
   | -- | Legacy shelley-based Commands
