+++ cardano-cli/src/Cardano/CLI/EraBased/Commands.hs
@@ -13,11 +13,11 @@ where
 
 import           Cardano.Api (ShelleyBasedEra (..), toCardanoEra)
 
+import           Cardano.CLI.Commands.Key
 import           Cardano.CLI.Commands.Node
 import           Cardano.CLI.Environment
 import           Cardano.CLI.EraBased.Commands.Address
 import           Cardano.CLI.EraBased.Commands.Genesis
-import           Cardano.CLI.EraBased.Commands.Key
 import           Cardano.CLI.EraBased.Commands.Query
 import           Cardano.CLI.EraBased.Commands.StakeAddress
 import           Cardano.CLI.EraBased.Commands.StakePool hiding (sbe)
@@ -27,12 +27,12 @@ import           Cardano.CLI.EraBased.Options.Address
 import           Cardano.CLI.EraBased.Options.Common
 import           Cardano.CLI.EraBased.Options.Genesis
 import           Cardano.CLI.EraBased.Options.Governance
-import           Cardano.CLI.EraBased.Options.Key
 import           Cardano.CLI.EraBased.Options.Query
 import           Cardano.CLI.EraBased.Options.StakeAddress
 import           Cardano.CLI.EraBased.Options.StakePool
 import           Cardano.CLI.EraBased.Options.TextView
 import           Cardano.CLI.EraBased.Options.Transaction
+import           Cardano.CLI.Options.Key
 import           Cardano.CLI.Options.Node
 
 import           Data.Foldable
