+++ cardano-cli/test/cardano-cli-test/Test/Cli/ITN.hs
@@ -8,7 +8,7 @@ module Test.Cli.ITN
   )
 where
 
-import           Cardano.CLI.EraBased.Run.Key (decodeBech32)
+import           Cardano.CLI.Run.Key (decodeBech32)
 
 import qualified Codec.Binary.Bech32 as Bech32
 import           Control.Monad (void)
